fix: use native scanout on all virtio-gpu variants
The useNativeScanout decision was keyed on VGA compatibility
(requiresNativeMode), which is false on virtio-vga-gl. This caused
enableController to skip the entire display pipeline (no
setupFramebufferResource, no refresh timer, no transferToHost2D)
and present a blue screen instead of a desktop.

The principled rule: native scanout is the rendering path whenever
the virtio-gpu control queue is functional, regardless of whether
a VGA BAR also exists. VGA compat only matters before a driver
loads; once this driver owns the device, the virtio-gpu protocol
is the only path that puts pixels on screen.

Also fixes a %zu format specifier on the common-map error log line
that was printing garbled values (1018u instead of 4120).

Variant status after this fix:
  virtio-gpu-gl-pci  — verified (desktop, cursor, mode switches)
  virtio-ramfb-gl    — verified (desktop, cursor, mode switches)
  virtio-vga-gl      — verified (desktop renders, blue screen gone)
  QXL                — verified (VMQemuVGA class, separate path)

+++ FB/VMVirtIOGPU.cpp
@@ -1087,8 +1087,8 @@ bool CLASS::initVirtIOGPU()
                     size_t required_common_size = common_offset + 24; // device_status (20) + 4 bytes
                     
                     if (common_map_size < required_common_size) {
-                        IOLog("VMVirtIOGPU: ERROR - Common map too small: %llu < %zu bytes\n", 
-                              (uint64_t)common_map_size, required_common_size);
+                        IOLog("VMVirtIOGPU: ERROR - Common map too small: %llu < %llu bytes\n",
+                              (uint64_t)common_map_size, (uint64_t)required_common_size);
                     } else {
                         volatile uint8_t* common_base = (volatile uint8_t*)common_map->getVirtualAddress();
                         if (common_base) {
